Struct moonbeam_extensions_evm::executor::wrapper::Config

pub struct Config {
    pub gas_ext_code: u64,
    pub gas_ext_code_hash: u64,
    pub gas_sstore_set: u64,
    pub gas_sstore_reset: u64,
    pub refund_sstore_clears: i64,
    pub gas_balance: u64,
    pub gas_sload: u64,
    pub gas_suicide: u64,
    pub gas_suicide_new_account: u64,
    pub gas_call: u64,
    pub gas_expbyte: u64,
    pub gas_transaction_create: u64,
    pub gas_transaction_call: u64,
    pub gas_transaction_zero_data: u64,
    pub gas_transaction_non_zero_data: u64,
    pub sstore_gas_metering: bool,
    pub sstore_revert_under_stipend: bool,
    pub err_on_call_with_more_gas: bool,
    pub call_l64_after_gas: bool,
    pub empty_considered_exists: bool,
    pub create_increase_nonce: bool,
    pub stack_limit: usize,
    pub memory_limit: usize,
    pub call_stack_limit: usize,
    pub create_contract_limit: Option<usize>,
    pub call_stipend: u64,
    pub has_delegate_call: bool,
    pub has_create2: bool,
    pub has_revert: bool,
    pub has_return_data: bool,
    pub has_bitwise_shifting: bool,
    pub has_chain_id: bool,
    pub has_self_balance: bool,
    pub has_ext_code_hash: bool,
    pub estimate: bool,
}

Runtime configuration.

Fields
gas_ext_code: u64

Gas paid for extcode.

gas_ext_code_hash: u64

Gas paid for extcodehash.

gas_sstore_set: u64

Gas paid for sstore set.

gas_sstore_reset: u64

Gas paid for sstore reset.

refund_sstore_clears: i64

Gas paid for sstore refund.

gas_balance: u64

Gas paid for BALANCE opcode.

gas_sload: u64

Gas paid for SLOAD opcode.

gas_suicide: u64

Gas paid for SUICIDE opcode.

gas_suicide_new_account: u64

Gas paid for SUICIDE opcode when it hits a new account.

gas_call: u64

Gas paid for CALL opcode.

gas_expbyte: u64

Gas paid for EXP opcode for every byte.

gas_transaction_create: u64

Gas paid for a contract creation transaction.

gas_transaction_call: u64

Gas paid for a message call transaction.

gas_transaction_zero_data: u64

Gas paid for zero data in a transaction.

gas_transaction_non_zero_data: u64

Gas paid for non-zero data in a transaction.

sstore_gas_metering: bool

EIP-1283.

sstore_revert_under_stipend: bool

EIP-1706.

err_on_call_with_more_gas: bool

Whether to throw out of gas error when CALL/CALLCODE/DELEGATECALL requires more than maximum amount of gas.

call_l64_after_gas: bool

Take l64 for callcreate after gas.

empty_considered_exists: bool

Whether empty account is considered exists.

create_increase_nonce: bool

Whether create transactions and create opcode increases nonce by one.

stack_limit: usize

Stack limit.

memory_limit: usize

Memory limit.

call_stack_limit: usize

Call limit.

create_contract_limit: Option<usize>

Create contract limit.

call_stipend: u64

Call stipend.

has_delegate_call: bool

Has delegate call.

has_create2: bool

Has create2.

has_revert: bool

Has revert.

has_return_data: bool

Has return data.

has_bitwise_shifting: bool

Has bitwise shifting.

has_chain_id: bool

Has chain ID.

has_self_balance: bool

Has self balance.

has_ext_code_hash: bool

Has ext code hash.

estimate: bool

Whether the gasometer is running in estimate mode.

Implementations

impl Config

pub const fn frontier() -> Config

Frontier hard fork configuration.

pub const fn istanbul() -> Config

Istanbul hard fork configuration.