Struct moonbeam_extensions_evm::executor::wrapper::StackExecutor

pub struct StackExecutor<'config, S> { /* fields omitted */ }

Stack-based executor.

Implementations

impl<'config, S> StackExecutor<'config, S> where
S: StackState<'config>,

pub fn new(state: S, config: &'config Config) -> StackExecutor<'config, S>

Create a new stack-based executor.

pub fn config(&self) -> &'config Config

Return a reference of the Config.

pub fn new_with_precompile(
state: S,
config: &'config Config,
precompile: fn(H160, &[u8], Option<u64>, &Context) -> Option<Result<(ExitSucceed, Vec<u8, Global>, u64), ExitError>>
) -> StackExecutor<'config, S>

Create a new stack-based executor with given precompiles.

pub fn state(&self) -> &S

pub fn state_mut(&mut self) -> &mut S

pub fn into_state(self) -> S

pub fn enter_substate(&mut self, gas_limit: u64, is_static: bool)

Create a substate executor from the current executor.

pub fn exit_substate(&mut self, kind: StackExitKind) -> Result<(), ExitError>

Exit a substate. Panic if it results an empty substate stack.

pub fn execute(&mut self, runtime: &mut Runtime<'_>) -> ExitReason

Execute the runtime until it returns.

pub fn gas(&self) -> u64

Get remaining gas.

pub fn transact_create(
&mut self,
caller: H160,
value: U256,
init_code: Vec<u8, Global>,
gas_limit: u64
) -> ExitReason

Execute a CREATE transaction.

pub fn transact_create2(
&mut self,
caller: H160,
value: U256,
init_code: Vec<u8, Global>,
salt: H256,
gas_limit: u64
) -> ExitReason

Execute a CREATE2 transaction.

pub fn transact_call(
&mut self,
caller: H160,
address: H160,
value: U256,
data: Vec<u8, Global>,
gas_limit: u64
) -> (ExitReason, Vec<u8, Global>)

Execute a CALL transaction.

pub fn used_gas(&self) -> u64

Get used gas for the current executor, given the price.

pub fn fee(&self, price: U256) -> U256

Get fee needed for the current executor, given the price.

pub fn nonce(&self, address: H160) -> U256

Get account nonce.

pub fn create_address(&self, scheme: CreateScheme) -> H160

Get the create address from given scheme.